work on reflection, non_null implementation
- start work on new reflection module, mostly stub implementations
- implement basic not nullable wrapper (optional wrapping a not nullable annotated pointer)
- renamed kb::scope -> kb::box

+++ KablunkEngine/include/Kablunk/Core/Memory/Memory.h
@@ -49,7 +49,7 @@ struct map_allocator_t
 			return p;
 
 		throw std::bad_alloc();
-		return (T*)0xDEADBEEF;
+		return (T*)static_cast<u64>(0xDEADBEEF);
 	}
 
 	void deallocate(T* p, std::size_t n) noexcept {
